# cmd/api/docs/swagger.yaml
basePath: /api
definitions:
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: okeke98@gmail.com
    name: For API Support
  description: 'Backend API built with Golang using Clean Code architecture. \nGithub:
    [https://github.com/okekefrancis112/ecommerce-gin].'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-commerce Application Backend API
paths:
...
  /account/wallet:
    get:
      description: API for user to get user wallet
      operationId: GetUserWallet
      responses:
        "200":
          description: Successfully retrieve user wallet
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve user wallet
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user wallet  (User)
      tags:
      - User Profile
  /account/wallet/transactions:
    get:
      description: API for user to get user wallet transaction
      operationId: GetUserWalletTransactions
      responses:
        "200":
          description: Successfully retrieved user wallet transactions
          schema:
            $ref: '#/definitions/response.Response'
        "204":
          description: No wallet transaction for user
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to retrieve user wallet transactions
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get user wallet  (User)
      tags:
      - User Profile
...
securityDefinitions:
  BearerAuth:
    description: 'Add prefix of Bearer before  token Ex: "Bearer token"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"